+++ lib/src/shcore.dart
@@ -31,7 +31,7 @@ final _shcore = DynamicLibrary.open('shcore.dll');
 int GetDpiForMonitor(
     int hMonitor, int dpiType, Pointer<Int32> dpiX, Pointer<Int32> dpiY) {
   final _GetDpiForMonitor = _shcore.lookupFunction<
-      IntPtr Function(IntPtr hMonitor, Int32 dpiType, Pointer<Int32> dpiX,
+      Int32 Function(IntPtr hMonitor, Int32 dpiType, Pointer<Int32> dpiX,
           Pointer<Int32> dpiY),
       int Function(int hMonitor, int dpiType, Pointer<Int32> dpiX,
           Pointer<Int32> dpiY)>('GetDpiForMonitor');
@@ -49,7 +49,7 @@ int GetDpiForMonitor(
 /// {@category shcore}
 int GetProcessDpiAwareness(int hprocess, Pointer<Int32> value) {
   final _GetProcessDpiAwareness = _shcore.lookupFunction<
-      IntPtr Function(IntPtr hprocess, Pointer<Int32> value),
+      Int32 Function(IntPtr hprocess, Pointer<Int32> value),
       int Function(
           int hprocess, Pointer<Int32> value)>('GetProcessDpiAwareness');
   return _GetProcessDpiAwareness(hprocess, value);
@@ -67,7 +67,7 @@ int GetProcessDpiAwareness(int hprocess, Pointer<Int32> value) {
 /// {@category shcore}
 int SetProcessDpiAwareness(int value) {
   final _SetProcessDpiAwareness = _shcore.lookupFunction<
-      IntPtr Function(Int32 value),
+      Int32 Function(Int32 value),
       int Function(int value)>('SetProcessDpiAwareness');
   return _SetProcessDpiAwareness(value);
 }
